+++ packages/amplify-category-api/src/__tests__/graphql-transformer/transform-graphql-schema-v2.test.ts
@@ -4,7 +4,7 @@ import { ApiCategoryFacade } from "@aws-amplify/amplify-cli-core";
 import { transformGraphQLSchemaV2 } from "../../graphql-transformer/transform-graphql-schema-v2";
 import { generateTransformerOptions } from "../../graphql-transformer/transformer-options-v2";
 import { getTransformerFactory } from "../../graphql-transformer/transformer-factory";
-import { getAppSyncAPIName } from "../../provider-utils/awscloudformation/utils/amplify-meta-utils";
+import { getAppSyncAPIName, getAccountConfig } from "../../provider-utils/awscloudformation/utils/amplify-meta-utils";
 
 jest.mock("@aws-amplify/amplify-cli-core");
 jest.mock("@aws-amplify/amplify-prompts");
@@ -19,6 +19,7 @@ describe("transformGraphQLSchemaV2", () => {
   const ApiCategoryFacadeMock = ApiCategoryFacade as jest.Mocked<typeof ApiCategoryFacade>;
   const generateTransformerOptionsMock = generateTransformerOptions as jest.Mock;
   const getAppSyncAPINameMock = getAppSyncAPIName as jest.Mock;
+  const getAccountConfigMock = getAccountConfig as jest.Mock;
 
   beforeEach(async () => {
     jest.clearAllMocks();
@@ -73,6 +74,10 @@ describe("transformGraphQLSchemaV2", () => {
       },
     });
     getAppSyncAPINameMock.mockReturnValue(["testapi"]);
+    getAccountConfigMock.mockReturnValue({
+      account: "account",
+      region: "region",
+    });
 
     await transformGraphQLSchemaV2(contextMock, {});
     expect(printerMock.warn).toBeCalledWith(
