+++ packages/amplify-graphql-transformer-core/src/transformer-context/stack-manager.ts
@@ -24,7 +24,6 @@ export class StackManager implements StackManagerProvider {
   constructor(app: App, resourceMapping: ResourceToStackMap, private accountConfig?: AccountConfig) {
     this.rootStack = new TransformerRootStack(app, 'transformer-root-stack', {
       synthesizer: this.stackSynthesizer,
-      env: { account: accountConfig?.accountId, region: accountConfig?.region },
     });
     // add Env Parameter to ensure to adhere to contract
     this.resourceToStackMap = new Map(Object.entries(resourceMapping));
